feat: integration with api
